# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel AI sentiment intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Sentiment = typeof Sentiment[keyof typeof Sentiment];


export const Sentiment = {
  positive: 'positive',
  negative: 'negative',
  neutral: 'neutral',
} as const;

export type Model = typeof Model[keyof typeof Model];


export const Model = {
  'logistic-regression': 'logistic-regression',
  'naive-bayes': 'naive-bayes',
  svm: 'svm',
  vader: 'vader',
  textblob: 'textblob',
} as const;

export interface SentimentCounts {
  total: number;
  positive: number;
  negative: number;
  neutral: number;
  positivePct: number;
  negativePct: number;
  neutralPct: number;
}

export interface TrendPoint {
  label: string;
  positive: number;
  negative: number;
  neutral: number;
}

export interface PlatformCount {
  platform: string;
  positive: number;
  negative: number;
  neutral: number;
}

export interface Dashboard {
  counts: SentimentCounts;
  trend: TrendPoint[];
  platforms: PlatformCount[];
  healthScore: number;
  analyzedToday: number;
  activeModels: number;
}

export type SignalType = typeof SignalType[keyof typeof SignalType];


export const SignalType = {
  spike: 'spike',
  cluster: 'cluster',
  insight: 'insight',
} as const;

export type SignalSeverity = typeof SignalSeverity[keyof typeof SignalSeverity];


export const SignalSeverity = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Signal {
  id: string;
  type: SignalType;
  title: string;
  detail: string;
  platform: string;
  sentiment: Sentiment;
  time: string;
  severity: SignalSeverity;
}

export interface AnalyzeInput {
  /**
     * @minLength 1
     * @maxLength 10000
     */
  text: string;
  model: Model;
}

export interface Keyword {
  word: string;
  count: number;
  sentiment: Sentiment;
}

export interface AnalysisResult {
  sentiment: Sentiment;
  confidence: number;
  polarity: number;
  model: Model;
  keywords: Keyword[];
  explanation: string;
}

export interface UploadInput {
  fileName: string;
  fileType: string;
  /** UTF-8 text or base64-encoded file content */
  content: string;
}

export interface UploadRow {
  id: string;
  text: string;
  platform: string;
  sentiment: Sentiment;
  confidence: number;
}

export interface UploadResult {
  fileName: string;
  totalRows: number;
  processedRows: number;
  rows: UploadRow[];
  summary: SentimentCounts;
}

export interface Benchmark {
  model: Model;
  name: string;
  family: string;
  accuracy: number;
  precision: number;
  recall: number;
  f1: number;
}

export type AnalyticsHighlights = {
  positiveTopic: string;
  negativeTopic: string;
  trendingKeyword: string;
};

export interface Analytics {
  keywords: Keyword[];
  highlights: AnalyticsHighlights;
  totalKeywords: number;
}