# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sentinel AI sentiment intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analyze:
    post:
      operationId: analyzeText
      tags: [intelligence]
      summary: Classify one social media text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeInput"
      responses:
        "200":
          description: Classification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Invalid input
  /upload:
    post:
      operationId: uploadDataset
      tags: [intelligence]
      summary: Parse and classify an uploaded dataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "200":
          description: Batch classification preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          description: Invalid upload
...
components:
  schemas:
...
    Sentiment:
      type: string
      enum: [positive, negative, neutral]
    Model:
      type: string
      enum: [logistic-regression, naive-bayes, svm, vader, textblob]
    SentimentCounts:
      type: object
      properties:
        total:
          type: number
        positive:
          type: number
        negative:
          type: number
        neutral:
          type: number
        positivePct:
          type: number
        negativePct:
          type: number
        neutralPct:
          type: number
      required: [total, positive, negative, neutral, positivePct, negativePct, neutralPct]
...
    AnalyzeInput:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 10000
        model:
          $ref: "#/components/schemas/Model"
      required: [text, model]
    Keyword:
      type: object
      properties:
        word:
          type: string
        count:
          type: number
        sentiment:
          $ref: "#/components/schemas/Sentiment"
      required: [word, count, sentiment]
    AnalysisResult:
      type: object
      properties:
        sentiment:
          $ref: "#/components/schemas/Sentiment"
        confidence:
          type: number
        polarity:
          type: number
        model:
          $ref: "#/components/schemas/Model"
        keywords:
          type: array
          items:
            $ref: "#/components/schemas/Keyword"
        explanation:
          type: string
      required: [sentiment, confidence, polarity, model, keywords, explanation]
    UploadInput:
      type: object
      properties:
        fileName:
          type: string
        fileType:
          type: string
        content:
          type: string
          description: UTF-8 text or base64-encoded file content
      required: [fileName, fileType, content]
    UploadRow:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        platform:
          type: string
        sentiment:
          $ref: "#/components/schemas/Sentiment"
        confidence:
          type: number
      required: [id, text, platform, sentiment, confidence]
    UploadResult:
      type: object
      properties:
        fileName:
          type: string
        totalRows:
          type: number
        processedRows:
          type: number
        rows:
          type: array
          items:
            $ref: "#/components/schemas/UploadRow"
        summary:
          $ref: "#/components/schemas/SentimentCounts"
      required: [fileName, totalRows, processedRows, rows, summary]
...