# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel AI sentiment intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { Keyword } from './keyword';
import type { Model } from './model';
import type { Sentiment } from './sentiment';

export interface AnalysisResult {
  sentiment: Sentiment;
  confidence: number;
  polarity: number;
  model: Model;
  keywords: Keyword[];
  explanation: string;
}