# lib/api-zod/src/generated/types/analyzeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel AI sentiment intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { Model } from './model';

export interface AnalyzeInput {
  /**
     * @minLength 1
     * @maxLength 10000
     */
  text: string;
  model: Model;
}