# lib/api-zod/src/generated/types/keyword.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel AI sentiment intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { Sentiment } from './sentiment';

export interface Keyword {
  word: string;
  count: number;
  sentiment: Sentiment;
}