# lib/api-zod/src/generated/types/model.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel AI sentiment intelligence API
 * OpenAPI spec version: 0.1.0
 */

export type Model = typeof Model[keyof typeof Model];


export const Model = {
  'logistic-regression': 'logistic-regression',
  'naive-bayes': 'naive-bayes',
  svm: 'svm',
  vader: 'vader',
  textblob: 'textblob',
} as const;