# lib/api-zod/src/generated/types/sentiment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel AI sentiment intelligence API
 * OpenAPI spec version: 0.1.0
 */

export type Sentiment = typeof Sentiment[keyof typeof Sentiment];


export const Sentiment = {
  positive: 'positive',
  negative: 'negative',
  neutral: 'neutral',
} as const;