# lib/api-zod/src/generated/types/uploadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel AI sentiment intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadInput {
  fileName: string;
  fileType: string;
  /** UTF-8 text or base64-encoded file content */
  content: string;
}