# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel AI sentiment intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { SentimentCounts } from './sentimentCounts';
import type { UploadRow } from './uploadRow';

export interface UploadResult {
  fileName: string;
  totalRows: number;
  processedRows: number;
  rows: UploadRow[];
  summary: SentimentCounts;
}