# lib/api-zod/src/generated/types/uploadRow.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel AI sentiment intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { Sentiment } from './sentiment';

export interface UploadRow {
  id: string;
  text: string;
  platform: string;
  sentiment: Sentiment;
  confidence: number;
}